requirements.txt
# --- Core ---
streamlit==1.33.0
pandas==2.2.1
numpy==1.26.4

# --- File Parsing ---
pypdf==4.2.0
python-docx==1.1.0

# --- Plots / Visualization ---
altair==5.3.0
plotly==5.20.0
seaborn==0.13.2
matplotlib==3.8.3

# --- AI Agents & APIs ---
openai==1.14.2
google-generativeai==0.5.2

# --- Utilities ---
python-dotenv==1.0.1

# --- Required for your project structure ---
# petri / risklens internal modules reference

